Finalised Kernel with printing
